@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "revanced-releases-api"
-version = "0.1.0"
+version = "1.1.0"
 description = "JSON API for ReVanced Releases"
 authors = ["Alexandre Teles <alexandre.teles@ufba.br>"]
 license = "AGPLv3"
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.11"
 fastapi = ">=0.85.0"
 httpx = {version = ">=0.23.0", extras = ["http2"]}
 httpx-cache = ">=0.6.0"
@@ -18,7 +18,7 @@ redis = ">=4.3.4"
 loguru = ">=0.6.0"
 sentry-sdk = ">=1.9.8"
 argon2-cffi = ">=21.3.0"
-hypercorn = {extras = ["uvloop"], version = ">=0.14.3"}
+uvloop = ">=0.17.0"
 cytoolz = ">=0.12.0"
 fastapi-paseto-auth = "^0.6.0"
 ujson = ">=5.5.0"
